requirements.txt
flask==3.0.0
elevenlabs==1.14.0
python-dotenv==1.0.0
openai==1.55.0
